+++ src/crc.c
@@ -40,7 +40,7 @@
   /*                                                                        */
   /*  --------------------------------------------------------------------  */
 
-static unsigned long crc32_tab[] = {
+static unsigned int crc32_tab[] = {
       0x00000000L, 0x77073096L, 0xee0e612cL, 0x990951baL, 0x076dc419L,
       0x706af48fL, 0xe963a535L, 0x9e6495a3L, 0x0edb8832L, 0x79dcb8a4L,
       0xe0d5e91eL, 0x97d2d988L, 0x09b64c2bL, 0x7eb17cbdL, 0xe7b82d07L,
@@ -97,11 +97,11 @@ static unsigned long crc32_tab[] = {
 
 /* Return a 32-bit CRC of the contents of the buffer. */
 
-unsigned long crc32(const unsigned char *s, unsigned int len, int salt)
+unsigned int crc32(const unsigned char *s, unsigned int len, int salt)
 {
-  unsigned long crc32val;
-  unsigned int  i;
-  
+  unsigned int crc32val;
+  unsigned int i;
+
   crc32val = salt;
   for (i = 0;  i < len;  i ++)
     {
